Moved ChoiceEnums & co. to enums.py
